requirements.txt
streamlit>=1.35.0
sentence-transformers>=2.7.0
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
pdfminer.six>=20231228
python-docx>=1.1.2
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
yake>=0.4.8
spacy>=3.7.4
rapidfuzz>=3.9.0
pyyaml>=6.0.1